pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Loginov <itsmeegorloginov@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" }
]

[tool.poetry.dependencies]
python = "^3.11"
coverage = "^7.2.7"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
